finish debug and audit rework
