+++ web/src/components/FileBar.vue
@@ -25,10 +25,10 @@
               <q-item clickable v-close-popup @click="showUploadMesh = true">
                 <q-item-section>Upload MeshAgent</q-item-section>
               </q-item>
-              <q-item clickable v-close-popup @click="showAuditManager = true">
+              <q-item clickable v-close-popup @click="showAuditManager">
                 <q-item-section>Audit Log</q-item-section>
               </q-item>
-              <q-item clickable v-close-popup @click="showDebugLog()">
+              <q-item clickable v-close-popup @click="showDebugLog">
                 <q-item-section>Debug Log</q-item-section>
               </q-item>
             </q-list>
@@ -154,12 +154,6 @@
           <PendingActions @close="showPendingActions = false" />
         </q-dialog>
       </div>
-      <!-- audit manager -->
-      <div class="q-pa-md q-gutter-sm">
-        <q-dialog v-model="showAuditManager" maximized transition-show="slide-up" transition-hide="slide-down">
-          <AuditManager @close="showAuditManager = false" />
-        </q-dialog>
-      </div>
       <!-- Install Agents -->
       <div class="q-pa-md q-gutter-sm">
         <q-dialog v-model="showInstallAgent">
@@ -223,7 +217,7 @@ import AutomationManager from "@/components/automation/AutomationManager";
 import AdminManager from "@/components/AdminManager";
 import InstallAgent from "@/components/modals/agents/InstallAgent";
 import UploadMesh from "@/components/modals/core/UploadMesh";
-import AuditManager from "@/components/AuditManager";
+import AuditManager from "@/components/logs/AuditManager";
 import BulkAction from "@/components/modals/agents/BulkAction";
 import Deployment from "@/components/Deployment";
 import ServerMaintenance from "@/components/modals/core/ServerMaintenance";
@@ -241,7 +235,6 @@ export default {
     InstallAgent,
     UploadMesh,
     AdminManager,
-    AuditManager,
     BulkAction,
     Deployment,
     ServerMaintenance,
@@ -256,7 +249,6 @@ export default {
       showAdminManager: false,
       showInstallAgent: false,
       showUploadMesh: false,
-      showAuditManager: false,
       showBulkAction: false,
       showPendingActions: false,
       bulkMode: null,
@@ -325,6 +317,23 @@ export default {
     showPermissionsManager() {
       this.$q.dialog({
         component: PermissionsManager,
+      });
+    },
+    showAuditManager() {
+      this.$q.dialog({
+        component: DialogWrapper,
+        componentProps: {
+          vuecomponent: AuditManager,
+          noCard: true,
+          componentProps: {
+            modal: true,
+          },
+          dialogProps: {
+            maximized: true,
+            ["transition-show"]: "slide-up",
+            ["transition-hide"]: "slide-down",
+          },
+        },
       });
     },
     showDebugLog() {
